// src/lib/types/contract.d.ts
export interface components {
    schemas: {
        ContractView: {
            /** Format: uuid */
            contractId?: string;
            /** @enum {string} */
            status?: "DRAFT" | "PENDING_SIGN" | "SIGNED" | "CANCELLED";
            /** Format: uuid */
            bookingId?: string;
            /** Format: uuid */
            unitId?: string;
            /** Format: uuid */
            customerId?: string;
            /** Format: uuid */
            buyerId?: string | null;
            /** Format: uuid */
            sellerId?: string | null;
            /** Format: uuid */
            templateId?: string;
            /** Format: date-time */
            createdAt?: string;
            draft?: {
                /** Format: uuid */
                draftId?: string;
                fileUrl?: string;
                /** Format: uuid */
                templateId?: string;
                /** Format: date-time */
                draftedAt?: string;
            } | null;
        };
        BookingConfirmedEvent: {
            /** Format: uuid */
            bookingId: string;
            /** Format: uuid */
            unitId: string;
            /** Format: uuid */
            customerId: string;
        };
        ErrorResponse: {
            error?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}